[Overflow-487] [Bugfix] The new roles are not displayed immediately on the users page when edit a role (Admin) (#286)
* [Overflow-487] [Bugfix] The new roles are not displayed immediately on the users page when edit a role (Admin)

* fixed delete role not updating user role selection

* switched manage/role user fetch to network-only

* added delete protection to original roles

+++ backend/app/GraphQL/Mutations/DeleteRole.php
@@ -2,7 +2,9 @@
 
 namespace App\GraphQL\Mutations;
 
+use App\Exceptions\CustomException;
 use App\Models\Role;
+use App\Models\User;
 
 final class DeleteRole
 {
@@ -12,9 +14,12 @@ final class DeleteRole
      */
     public function __invoke($_, array $args)
     {
+        if ($args['id'] <= 3 && $args['id'] >= 1) {
+            return new CustomException('Cannot delete original roles');
+        }
         $role = Role::find($args['id']);
+        User::where('role_id', $args['id'])->update(['role_id' => 3]);
         $role->delete();
-
         return $role;
     }
 }
